optimized donuts generation by reusage of objects
